requirements.txt
Flask
python-dotenv
psycopg2-binary
Flask-WTF
trycourier
email_validator
PyJWT